Profile Image Upload (#194)
Changes Introduced:
Gallery Integration: The profile settings bottom sheet modal now enables users to access their device's gallery, upon clicking on their current profile images, facilitating the selection and upload of a new profile image.
Firebase Storage Integration: Upon selecting an image from the device's gallery, it is uploaded to Firebase Storage. The stored image is given a filename of the user's ID, ensuring uniqueness and ease of retrieval.
Data Synchronization: The uploaded image is associated with the user's profile and is synchronized across the application. Any updates to the profile image are reflected on all screens that display the user's data, such as team and profile screens.