Gehrkej/Reset season (#236)
A coach can reset a season which will clear all the leaderboards. This is just setting all fields in each document in the "best_attempts" collection to null. Adjustments were made to the all time record as well in order for it to display even when the leaderboard was wiped.